pyproject.toml
[tool.poetry]
name = "langton-ant"
version = "0.1.0"
description = ""
authors = ["DorLex <dorozhinskiysanya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.26.1"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
